major stylistic changes across the board (no pun intended)
